Change PoW from nonce search to Unprll hash chain generation
The "Unprll PoW", for lack of a better name.

It operates by starting with base data, obtaining its hash, checking the
hash against the current difficulty, then rinse and repeat using the
generated hash as input to the above.

This is non-parallelizable (namesake) due to the fact that each hash
depends on its previous hash.